Adding a dataset available to end user
why: to provide a simple dataset available in the package for end user to test and practice how to use the package.

how: creating a data-raw folder with DATASET.R formating my dataset, and make it available for end user in data folder. Adding documentation with checkhelper (doc_my_dataset.R)

Issue #